+++ CHANGELOG.md
@@ -2,11 +2,19 @@
 
 ## develop
 
+### Breaking changes
+
+- BREAKING(task): drop support for `multilabel` training in `SpeakerDiarization` task
+- BREAKING(task): drop support for `warm_up` option in `SpeakerDiarization` task
+- BREAKING(task): drop support for `weigh_by_cardinality` option in `SpeakerDiarization` task
+- BREAKING(task): drop support for `vad_loss` option in `SpeakerDiarization` task
+
 ### New features
 
-- feat: add support for `k-means` clustering
-- feat: add `"hidden"` option to `ProgressHook`
-- feat: add `FilterByNumberOfSpeakers` protocol files filter
+- feat(clustering): add support for `k-means` clustering
+- feat(task): add support for manual optimization in `SpeakerDiarization` task
+- feat(utils): add `hidden` option to `ProgressHook`
+- feat(utils): add `FilterByNumberOfSpeakers` protocol files filter
 
 ### Fixes